Re-create project structure to build with .NET 4.8/the `dotnet` CLI
The existing .sln and .csproj files did not load correctly in msbuild. And from past experience, it's easier to create project files from scratch when porting things over to the new non-Visual-Studio tooling.

The new project layout also means that `AssemblyInfo.cs` needs to be removed (this code is now generated by the build based on properties in .csproj file).

One necessary code change was to implement the `if(InvokeRequired) Invoke` pattern for `MainFrame.generationFinished` method. Manipulating the GUI from a background thread is not allowed and leads to an exception in .NET 4.8. It's possible that there are other cases where this pattern needs to be applied.

Also:
 - Remove unused and empty GridPanel.resx and MainFrame.resx
 - Use `ResourceManager` instead of `ResXResourceReader` to read Messages.xml (which is now renamed to Messages.resx so that it gets converted into embedded resources automatically)
 - Address warnings in `ErrorMessages` (make "enum values" `readonly`, use `string` instead of `String`)
 - Read version number from the assembly name (instead of hard-coding it).

Issue #1